+++ marshmallow/__init__.py
@@ -11,7 +11,7 @@ from marshmallow.decorators import (
     pre_dump, post_dump, pre_load, post_load, validates, validates_schema
 )
 from marshmallow.utils import pprint, missing
-from marshmallow.exceptions import MarshallingError, UnmarshallingError, ValidationError
+from marshmallow.exceptions import ValidationError
 
 __version__ = '2.0.0'
 __author__ = 'Steven Loria'
@@ -29,8 +29,6 @@ __all__ = [
     'pprint',
     'MarshalResult',
     'UnmarshalResult',
-    'MarshallingError',
-    'UnmarshallingError',
     'ValidationError',
     'missing',
 ]