+++ marshmallow/exceptions.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 """Exception classes for marshmallow-related errors."""
-import warnings
 
 from marshmallow.compat import basestring
 
@@ -47,31 +46,3 @@ class RegistryError(NameError):
     class registry.
     """
     pass
-
-
-class MarshallingError(ValidationError):
-    """Raised in case of a marshalling error. If raised during serialization,
-    the error is caught and the error message is stored in an ``errors``
-    dictionary (unless ``strict`` mode is turned on).
-
-    .. deprecated:: 2.0.0
-        Use :exc:`ValidationError` instead.
-    """
-    def __init__(self, *args, **kwargs):
-        warnings.warn('MarshallingError is deprecated. Raise a ValidationError instead',
-                      category=DeprecationWarning)
-        super(MarshallingError, self).__init__(*args, **kwargs)
-
-
-class UnmarshallingError(ValidationError):
-    """Raised when invalid data are passed to a deserialization function. If
-    raised during deserialization, the error is caught and the error message
-    is stored in an ``errors`` dictionary.
-
-    .. deprecated:: 2.0.0
-        Use :exc:`ValidationError` instead.
-    """
-    def __init__(self, *args, **kwargs):
-        warnings.warn('UnmarshallingError is deprecated. Raise a ValidationError instead',
-                      category=DeprecationWarning)
-        super(UnmarshallingError, self).__init__(*args, **kwargs)