+++ tests/test_exceptions.py
@@ -1,8 +1,5 @@
 # -*- coding: utf-8 -*-
-import pytest
-
-from marshmallow.exceptions import ValidationError, MarshallingError, UnmarshallingError
-from marshmallow import fields, Schema
+from marshmallow.exceptions import ValidationError
 
 
 class TestValidationError:
@@ -32,55 +29,3 @@ class TestValidationError:
 
         err2 = ValidationError('invalid email', 'email')
         assert str(err2) == 'invalid email'
-
-
-class TestMarshallingError:
-
-    def test_deprecated(self):
-        pytest.deprecated_call(MarshallingError, 'foo')
-
-    def test_can_store_field_and_field_name(self):
-        field_name = 'foo'
-        field = fields.Str()
-        err = MarshallingError('something went wrong', fields=[field],
-                               field_names=[field_name])
-        assert err.fields == [field]
-        assert err.field_names == [field_name]
-
-    def test_can_be_raised_by_custom_field(self):
-        class MyField(fields.Field):
-            def _serialize(self, val, attr, obj):
-                raise MarshallingError('oops')
-
-        class MySchema(Schema):
-            foo = MyField()
-
-        s = MySchema()
-        result = s.dump({'foo': 42})
-        assert 'foo' in result.errors
-        assert result.errors['foo'] == ['oops']
-
-class TestUnmarshallingError:
-
-    def test_deprecated(self):
-        pytest.deprecated_call(UnmarshallingError, 'foo')
-
-    def test_can_store_field_and_field_name(self):
-        field_name = 'foo'
-        field = fields.Str()
-        err = UnmarshallingError('something went wrong', fields=[field],
-                                 field_names=[field_name])
-        assert err.fields == [field]
-        assert err.field_names == [field_name]
-
-    def test_can_be_raised_by_validator(self):
-        def validator(val):
-            raise UnmarshallingError('oops')
-
-        class MySchema(Schema):
-            foo = fields.Field(validate=[validator])
-
-        s = MySchema()
-        result = s.load({'foo': 42})
-        assert 'foo' in result.errors
-        assert result.errors['foo'] == ['oops']